# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Property api Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://laraveldecouverte.test:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/properties": {
            "get": {
                "tags": [
                    "GET"
                ],
                "summary": "Récupéré la liste des propriétés",
                "description": "Retourne la liste complète de toute les propriétés.",
                "operationId": "getAllProperties",
                "responses": {
                    "200": {
                        "description": "Opération réussis",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "400": {
                        "description": "Requête erroné"
                    },
                    "404": {
                        "description": "Page introuvable"
                    }
                }
            }
        },
        "/property/create": {
            "post": {
                "tags": [
                    "POST"
                ],
                "summary": "Crée une propriété",
                "description": "Retourne une propriété.",
                "operationId": "createProperty",
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "image",
                                    "postcode",
                                    "city",
                                    "address",
                                    "room",
                                    "price",
                                    "size",
                                    "floor"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Titre de la propriété",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "l'image de la propriété",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "le prix de la propriété",
                                        "type": "number"
                                    },
                                    "size": {
                                        "description": "Surface de la propriété",
                                        "type": "number"
                                    },
                                    "floor": {
                                        "description": "Nombre d'étage de la propriété",
                                        "type": "number"
                                    },
                                    "description": {
                                        "description": "Description de la propriété",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "L'address de la propriété",
                                        "type": "string"
                                    },
                                    "postcode": {
                                        "description": "Code postale de la propriété",
                                        "type": "number"
                                    },
                                    "room": {
                                        "description": "Nombre de pièces de la propriété",
                                        "type": "number"
                                    },
                                    "city": {
                                        "description": "Ville de la propriété",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Opération réussis",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "400": {
                        "description": "Requête erroné"
                    },
                    "404": {
                        "description": "Page introuvable"
                    }
                }
            }
        },
        "/property/{id}": {
            "get": {
                "tags": [
                    "GET"
                ],
                "summary": "Récupéré une propriété",
                "description": "Retourne une propriété.",
                "operationId": "getProperty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "l'id de la propriété",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Opération réussis",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "400": {
                        "description": "Requête erroné"
                    },
                    "404": {
                        "description": "Page introuvable"
                    }
                }
            }
        },
        "/property/update/{id}": {
            "put": {
                "tags": [
                    "PUT"
                ],
                "summary": "Met à jour une propriété",
                "description": "Modifie une propriété.",
                "operationId": "updateProperty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "l'id de la propriété",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Input data format",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "image",
                                    "postcode",
                                    "city",
                                    "address",
                                    "room",
                                    "price",
                                    "size",
                                    "floor"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Titre de la propriété",
                                        "type": "string"
                                    },
                                    "image": {
                                        "description": "l'image de la propriété",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "le prix de la propriété",
                                        "type": "number"
                                    },
                                    "size": {
                                        "description": "Surface de la propriété",
                                        "type": "number"
                                    },
                                    "floor": {
                                        "description": "Nombre d'étage de la propriété",
                                        "type": "number"
                                    },
                                    "description": {
                                        "description": "Description de la propriété",
                                        "type": "string"
                                    },
                                    "address": {
                                        "description": "L'address de la propriété",
                                        "type": "string"
                                    },
                                    "postcode": {
                                        "description": "Code postale de la propriété",
                                        "type": "number"
                                    },
                                    "room": {
                                        "description": "Nombre de pièces de la propriété",
                                        "type": "number"
                                    },
                                    "city": {
                                        "description": "Ville de la propriété",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Opération réussis",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non authentifié"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "400": {
                        "description": "Requête erroné"
                    },
                    "404": {
                        "description": "Page introuvable"
                    }
                }
            }
        },
        "/property/delete/{id}": {
            "delete": {
                "tags": [
                    "DELETE"
                ],
                "summary": "Supprimé une propriété",
                "description": "Supprimé une propriété",
                "operationId": "deleteProperty",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id de la propriété",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalide id"
                    },
                    "404": {
                        "description": "Propriété non trouvé"
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}